Initialized new branch on multi-dimensional interpolation: Initial changes to make new interface class for multi-dimensional interpolation, such that more interpolators can be added
